+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release: dd.mm.yyyy
 
 ### New
 
+- New app icon by [Parakeet](https://parakeet.co) ([#286](https://github.com/kasper/phoenix/issues/286))!
 - Improved documentation! 👩🏼‍💻 The documentation is now powered by Docusaurus and hosted at https://kasper.github.io/phoenix/.
 - Add support for an input modal 🚀 ([#163](https://github.com/kasper/phoenix/issues/163)).
 - Add support to customise text alignment, font and colour in modal ([#158](https://github.com/kasper/phoenix/issues/158), [#242](https://github.com/kasper/phoenix/issues/242)).